\name{clean.numeric}
\alias{clean.numeric}
\title{Replace user-specified "missing data" codes with NA and convert to numeric.}
\usage{clean.numeric(x, missing=list("---"))}
\description{Replace user-specified "missing data" codes with NA and convert to numeric.}
\value{a numeric array}
\arguments{\item{x}{the array to clean}
\item{missing}{a list of codes considered equivalent to NA}}
