\name{metfiles-package}
\title{Reads meteorology data from formats commonly used in dispersion modeling}
\description{Reads meteorology data from formats commonly used in dispersion modeling.}
\details{\tabular{ll}{
Package: \tab metfiles\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2011-06-02\cr
License: \tab BSD\cr
LazyLoad: \tab yes\cr
}

Supported formats include SCRAM surface (aka MET144), SCRAM upper air, and SAMSON hourly.}
\alias{metfiles}
\docType{package}
\author{David Holstius}
\references{\url{http://www.epa.gov/scram001/surface/surfor.txt}}
\keyword{package}
\seealso{\code{\link{read.SCRAM.surface}}}
