\name{read.SAMSON}
\alias{read.SAMSON}
\title{Read SAMSON meteorological data.}
\usage{read.SAMSON(file, ...)}
\description{Read SAMSON meteorological data.}
\value{a data.frame with the data contained in the file}
\arguments{\item{file}{the file to read}
\item{\dots}{further arguments to be passed to \code{read.fortran()}}}
