\name{read.SCRAM.upper}
\alias{read.SCRAM.upper}
\title{Read SCRAM upper-air meteorological data.}
\usage{read.SCRAM.upper(file, ...)}
\description{Read SCRAM upper-air meteorological data.}
\value{a data.frame with the data contained in the file}
\seealso{\code{\link{read.SCRAM.surface}}}
\arguments{\item{file}{the file to read}
\item{\dots}{further arguments to be passed to \code{read.fwf()}}}
