\name{replace.all}
\alias{replace.all}
\title{replace...}
\usage{replace.all(x, what, value)}
\description{\code{replace.all} is a wrapper around \code{replace}.}
\value{a vector with the values replaced.}
\arguments{\item{x}{vector}
\item{what}{list of values to replace (usually one or more character strings)}
\item{replacement}{value}}
