\name{read.SCRAM.surface}
\alias{read.SCRAM.surface}
\title{Read surface meteorological data from a file.}
\usage{read.SCRAM.surface(file, ...)}
\description{Read surface meteorological data from a file.}
\value{a data.frame with the data contained in the file}
\seealso{\code{\link{read.SCRAM.upper}}}
\references{\url{http://www.epa.gov/scram001/surface/surfor.txt}}
\arguments{\item{file}{the file to read}
\item{\dots}{further arguments to be passed to \code{read.fwf()}}}
