\name{read.SCRAM.upper}
\alias{read.SCRAM.upper}
\title{Read upper-air meteorological data from a file.}
\usage{read.SCRAM.upper(path, ...)}
\description{Read upper-air meteorological data from a file.}
\value{a data.frame with the data contained in the path}
\seealso{\code{\link{read.SCRAM.surface}}}
\arguments{\item{path}{the file to read}
\item{\dots}{further arguments to be passed to \code{read.fwf()}}}
